# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "https://archimed.justcode.am/",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/get_basket": {
            "get": {
                "tags": [
                    "Basket"
                ],
                "summary": "get_basket",
                "description": "Получаем список моей корзини если нужен один экзепляр отправлаем  /api/get_basket/2 ОТПРАВЛАЕМ TOKEN",
                "operationId": "get_basket",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Basket data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/doctor_visit_time": {
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "doctor_visit_time",
                "description": "Получаем Список свободных часов Одного доктора для конкретной даты",
                "operationId": "doctor_visit_time",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subject_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "PL_EXAM_ID": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2023-05-18"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2023-05-18"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visit Time Data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong Data sended subject_id and PL_EXAM_ID no exist",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/add_new_order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Add a new order (online or offline)",
                "operationId": "b1c6ce63a2570e1d49d7047e7efcb95a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pay_type"
                                ],
                                "properties": {
                                    "pay_type": {
                                        "description": "Payment type (online or offline)",
                                        "type": "string",
                                        "enum": [
                                            "online Or offline"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order_payment_successfully": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Process successful order payment",
                "operationId": "orderPaymentSuccessfully",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "basket_id": {
                                        "type": "string",
                                        "example": "DEF456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order Created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or wrong order_id/basket_id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error or wrong order_id/basket_id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/update_basket_device": {
            "post": {
                "tags": [
                    "Basket"
                ],
                "summary": "Update basket device",
                "operationId": "1f9ab62c7b4fc9c75fd2db925b407a1e",
                "parameters": [
                    {
                        "name": "basket_id",
                        "in": "query",
                        "description": "Basket ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "day",
                        "in": "query",
                        "description": "Day",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_day",
                        "in": "query",
                        "description": "Start day",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end_day",
                        "in": "query",
                        "description": "End day",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": " "
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}